requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
Django==4.1.10
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
mongoengine==0.27.0
Pillow==10.0.0
pymongo==4.4.1
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
